requirements.txt
appdirs==1.4.3
argon2-cffi==19.1.0
bcrypt==3.1.7
certifi==2019.6.16
cfe==0.0.15
cffi==1.12.3
chardet==3.0.4
dj-database-url==0.5.0
django-crispy-forms==1.7.2
django-registration-redux==2.6
djangorestframework==3.9.4
gunicorn==19.9.0
idna==2.8
mysqlclient==1.4.2.post1
olefile==0.46
packaging==19.0
# Pillow==6.1.0
prettytable==0.7.2
psycopg2==2.8.3
pycparser==2.19
pyparsing==2.4.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.2
